refactor(camtools): move file extension validation functions from util.py to sanity.py for better logical organization docs: add API documentation for sanity and util modules
