+++ docs/api/sanity.rst
@@ -1,0 +1,9 @@
+ct.sanity
+=========
+
+.. currentmodule:: ct
+
+.. automodule:: ct.sanity
+   :members:
+   :undoc-members:
+   :show-inheritance:
