+++ docs/api/util.rst
@@ -1,0 +1,9 @@
+ct.util
+=======
+
+.. currentmodule:: ct
+
+.. automodule:: ct.util
+   :members:
+   :undoc-members:
+   :show-inheritance:
